Implemented GBuffer. Cleaned up a lot of the code.
fixed camera bug (the hyperspeed calculation i had applied was modifying the normal of the velocity and causing weird offsets)

removed preamble (for now)

+++ GenGein/GenGein/Source/Render/Objects/Components/Shape.cpp
@@ -30,7 +30,7 @@ void Shape::Create(const ShapeType a_type)
 	// #SECRET: Beefy boi returns a big chunk of dates
 	CreateShapeProperties(vertices, indices);
 
-	m_bufferInfo.m_IndexCount = indices.size();
+	m_bufferInfo.m_IndexCount = (unsigned int)indices.size();
 
 	unsigned int& vao = m_bufferInfo.m_VAO;
 	unsigned int& vbo = m_bufferInfo.m_VBO;
